refactor: move useAsyncGuard and useAutoRefresh to @tracepilot/ui package
- Move useAsyncGuard.ts and useAutoRefresh.ts from desktop to packages/ui
- Update 16 desktop files to import useAsyncGuard from @tracepilot/ui
- Update 5 desktop views to import useAutoRefresh from @tracepilot/ui
- Merge duplicate @tracepilot/ui imports across all affected files
- Fix 3 test files using vi.mock('@tracepilot/ui') to use importOriginal
  pattern (prevents blanking useAsyncGuard when mocking toErrorMessage etc.)
- Delete duplicate test files from desktop (canonical tests now in UI pkg)
- Derive AsyncGuardLike from AsyncGuard via Pick<> for type safety
- Add vite devDependency to UI package for import.meta.env types
- Export AsyncGuard, AsyncGuardToken, useAsyncGuard, AutoRefreshOptions,
  useAutoRefresh from composables barrel

Net: -1,138 lines from desktop, +304 lines in UI package (shared)
Tests: UI 741 passed, Desktop 1177 passed, typechecks clean
