+++ apps/desktop/src/__tests__/stores/mcp.test.ts
@@ -47,9 +47,13 @@ vi.mock("@/utils/logger", () => ({
 }));
 
 // ── Mock @tracepilot/ui ────────────────────────────────────────
-vi.mock("@tracepilot/ui", () => ({
-  toErrorMessage: (e: unknown) => (e instanceof Error ? e.message : String(e)),
-}));
+vi.mock("@tracepilot/ui", async (importOriginal) => {
+  const actual = await importOriginal<typeof import("@tracepilot/ui")>();
+  return {
+    ...actual,
+    toErrorMessage: (e: unknown) => (e instanceof Error ? e.message : String(e)),
+  };
+});
 
 // ── Fixtures ───────────────────────────────────────────────────
 const FIXTURE_CONFIG: McpServerConfig = {
