+++ apps/desktop/src/composables/useToolResultLoader.ts
@@ -1,6 +1,6 @@
 import { getToolResult } from "@tracepilot/client";
 import { reactive, watch } from "vue";
-import { useAsyncGuard } from "@/composables/useAsyncGuard";
+import { useAsyncGuard } from "@tracepilot/ui";
 import { formatObjectResult } from "@/utils/formatResult";
 import { logError } from "@/utils/logger";
