+++ apps/desktop/src/stores/mcp.ts
@@ -21,10 +21,9 @@ import type {
   McpSummary,
   McpTool,
 } from "@tracepilot/types";
-import { toErrorMessage } from "@tracepilot/ui";
+import { toErrorMessage, useAsyncGuard } from "@tracepilot/ui";
 import { defineStore } from "pinia";
 import { computed, ref } from "vue";
-import { useAsyncGuard } from "@/composables/useAsyncGuard";
 import { logWarn } from "@/utils/logger";
 
 export const useMcpStore = defineStore("mcp", () => {
