+++ packages/ui/package.json
@@ -34,6 +34,7 @@
     "@vue/test-utils": "^2.4.6",
     "jsdom": "^28.1.0",
     "typescript": "^5.8.0",
+    "vite": "^6.3.0",
     "vitest": "^3.2.0",
     "vue-router": "^4.6.4",
     "vue-tsc": "^2.2.0"
