+++ packages/ui/src/composables/index.ts
@@ -1,4 +1,11 @@
 export type {
+  AsyncGuard,
+  AsyncGuardToken,
+} from "./useAsyncGuard";
+export { useAsyncGuard } from "./useAsyncGuard";
+export type { AutoRefreshOptions } from "./useAutoRefresh";
+export { useAutoRefresh } from "./useAutoRefresh";
+export type {
   ChartTooltipState,
   UseChartTooltipReturn,
 } from "./useChartTooltip";
