+++ packages/ui/src/composables/useAutoRefresh.ts
